requirements.txt
pandas
numpy
streamlit
matplotlib
tensorflow
seaborn
scikit-learn
pillow

